Split links away from tmt.base.core (#4901)
