+++ tests/unit/test_base.py
@@ -11,7 +11,8 @@ import pytest
 import tmt
 import tmt.result
 import tmt.utils
-from tmt.base.core import FmfId, Link, LinkNeedle, Links, expand_node_data
+from tmt.base.core import FmfId, expand_node_data
+from tmt.base.links import Link, LinkNeedle, Links
 from tmt.utils import Path, SpecificationError
 
 if TYPE_CHECKING:
